setup.py: Make sqlite and redis dependencies optional
This makes it possible to run the daemon without the aiosqlite and aioredis
packages installed, as long as those backends are not used by the running
config.

The conditional imports were already there, but because the packages are
declared as dependencies in setup.py, the setuptools-generated entry point
will refuse to run if they are not present.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>