Run mta-sts-deamon as non-root
In order to improve the security of the image, droping the unneeded root
permissions is a good idea. This improves security of the image and
allows better handling of the container in high security environments.

This might break existing setups due to `/var/lib/mta-sts` being created
by root/with root permissions previously. This can be solved by removing
the volume and re-creating it.