Add security-option for additonal hardening
Since we already dropped a lot of capabilities, lel's make sure we never
get them back in the container.

This should make sure, that even in case of a process escaping a process
escaping the container, it should be unable to get root permissions,
since the container drops those on creation.